requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-JWT-Extended
pytest
pytest-flask
black
flake8
psycopg2-binary==2.9.10
python-dotenv==1.1.0
